third_party/blink/web_tests/external/wpt/wasm/core/memory64/memory64.wast.js-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] #1 Reinitialize the default imports
  promise_test: Unhandled rejection with value: object "TypeError: Cannot convert 10 to a BigInt"
[FAIL] #6 Test that WebAssembly compilation succeeds (memory64.wast:8)
  assert_equals: expected false but got true
[FAIL] #7 Test that WebAssembly compilation succeeds (memory64.wast:9)
  assert_equals: expected false but got true
[FAIL] #27 Reinitialize the default imports
  promise_test: Unhandled rejection with value: object "TypeError: Cannot convert 10 to a BigInt"
[FAIL] #28 Test that WebAssembly compilation succeeds (memory64.wast:4)
  assert_true: WebAssembly.compile failed unexpectedly with TypeError: Cannot convert 10 to a BigInt at {loc} expected true got false
[FAIL] #29 Test that WebAssembly instantiation succeeds
  assert_true: unexpected instantiation error, observed TypeError: Cannot read properties of undefined (reading 'source') \n    at instance (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:207:15)\n    at memory64_wast_js (http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:7:10)\n    at http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
[FAIL] #36 Test that WebAssembly compilation succeeds (memory64.wast:8)
  assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.compile(): initial memory size (281474976710656 pages) is larger than implementation limit (262144 pages) @+16 at {loc} expected true got false
[FAIL] #37 Test that WebAssembly compilation succeeds (memory64.wast:9)
  assert_true: WebAssembly.compile failed unexpectedly with CompileError: WebAssembly.compile(): maximum memory size (281474976710656 pages) is larger than implementation limit (262144 pages) @+17 at {loc} expected true got false
[FAIL] #38 Test that WebAssembly instantiation succeeds
  assert_true: unexpected instantiation error, observed TypeError: Cannot read properties of undefined (reading 'source') \n    at instance (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:207:15)\n    at memory64_wast_js (http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:34:10)\n    at http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
[FAIL] #66 Test that WebAssembly instantiation succeeds (wrapper)
  assert_true: unexpected instantiation error, observed TypeError: WebAssembly.instantiate(): Import #1 "spectest": module is not an object or function \n    at instance (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:207:15)\n    at http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:115:16\n    at run (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:249:31)\n    at memory64_wast_js (http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:115:1)\n    at http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
[FAIL] #67 run
  assert_true: unexpected runtime error, observed TypeError: Cannot read properties of undefined (reading 'run') \n    at run (http://web-platform.test:8001/wasm/core/js/harness/async_index.js:248:15)\n    at memory64_wast_js (http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:115:1)\n    at http://web-platform.test:8001/wasm/core/js/memory64/memory64.wast.js:237:3 expected true got false
Harness: the test ran to completion.
